Allow 2-tuple input as dict
FOr the 3-tuple input, that we interpret as subj-pred-obj

But how should we allow facts of N-ary att/value pairs? Simplest might be p : x, q: r BUT this means : and , are both reserved characters? Or would we use native durable_rules assertion statements anyway and not bother with the magic?